+++ Terra/src/utopia/terra/controller/coordinate/distance/VectorDistanceOps.scala
@@ -3,7 +3,21 @@ package utopia.terra.controller.coordinate.distance
 import utopia.flow.collection.immutable.Pair
 import utopia.flow.operator.{Combinable, HasLength, Reversible}
 import utopia.paradigm.measurement.Distance
+import utopia.terra.controller.coordinate.world.VectorDistanceConversion
 import utopia.terra.model.world.WorldPoint
+
+
+object VectorDistanceOps
+{
+	/**
+	  * Creates a new vector distance ops -instance using implicitly available scaling information
+	  * @param vw Implicit world view to assume
+	  * @tparam V Type of vectors compared
+	  * @return A new instance for calculating vector distances
+	  */
+	def implicitly[V <: HasLength with Reversible[V] with Combinable[V, V]](implicit vw: VectorDistanceConversion) =
+		new VectorDistanceOps[V](vw.unitDistance)
+}
 
 /**
  * Distance-calculation algorithm for calculating linear distance between two points using their vector forms.