+++ Terra/src/utopia/terra/controller/coordinate/world/WorldView.scala
@@ -1,8 +1,5 @@
 package utopia.terra.controller.coordinate.world
 
-import utopia.paradigm.measurement.Distance
-import utopia.terra.model.angular.LatLong
-
 /**
  * Common trait for different world representation models used for coordinate transformations
  * @author Mikko Hilpinen
@@ -13,27 +10,5 @@ import utopia.terra.model.angular.LatLong
  * @tparam Aerial Aerial (3D) point type
  */
 trait WorldView[-V2D, -V3D, +Surface, +Aerial]
-{
-	/**
-	 * @param latLong A latitude-longitude coordinate
-	 * @return A point on the Earth's surface that matches that coordinate
-	 */
-	def apply(latLong: LatLong): Surface
-	/**
-	 * @param latLong A latitude-longitude coordinate
-	 * @param altitude Altitude above the sea level
-	 * @return A point within the world that matches this location
-	 */
-	def apply(latLong: LatLong, altitude: Distance): Aerial
-	
-	/**
-	 * @param vector A vector in this world view system
-	 * @return A surface point that matches that vector
-	 */
-	def surfaceVector(vector: V2D): Surface
-	/**
-	 * @param vector A vector in this world view system
-	 * @return A 3D point that matches that vector
-	 */
-	def aerialVector(vector: V3D): Aerial
-}
+	extends LatLongToWorldPoint[Surface, Aerial] with VectorToWorldPoint[V2D, V3D, Surface, Aerial]
+		with VectorDistanceConversion