+++ Terra/src/utopia/terra/model/world/circle/CirclePoint.scala
@@ -2,13 +2,16 @@ package utopia.terra.model.world.circle
 
 import utopia.flow.operator.EqualsBy
 import utopia.paradigm.measurement.Distance
+import utopia.paradigm.shape.shape2d.vector.Vector2D
 import utopia.paradigm.shape.shape3d.Vector3D
 import utopia.terra.controller.coordinate.distance.{DistanceOps, VectorDistanceOps}
-import utopia.terra.controller.coordinate.world.CircleOfEarth
+import utopia.terra.controller.coordinate.world.{CircleOfEarth, LatLongToWorldPoint, VectorToWorldPoint}
 import utopia.terra.model.angular.LatLong
 import utopia.terra.model.world.AerialPoint
 
 object CirclePoint
+	extends VectorToWorldPoint[Vector2D, Vector3D, CircleSurfacePoint, CirclePoint]
+		with LatLongToWorldPoint[CircleSurfacePoint, CirclePoint]
 {
 	// ATTRIBUTES   ------------------------
 	
@@ -16,6 +19,16 @@ object CirclePoint
 	 * Algorithm used for calculating distances between two points of this kind
 	 */
 	implicit val distanceOps: DistanceOps[CirclePoint] = new VectorDistanceOps[Vector3D](CircleOfEarth.unitDistance)
+	
+	
+	// IMPLEMENTED  ------------------------
+	
+	override def apply(latLong: LatLong): CircleSurfacePoint = CircleSurfacePoint(latLong)
+	override def apply(latLong: LatLong, altitude: Distance): CirclePoint =
+		new _CirclePoint(CircleSurfacePoint(latLong), altitude)
+	
+	override def surfaceVector(vector: Vector2D): CircleSurfacePoint = CircleSurfacePoint(vector)
+	override def aerialVector(vector: Vector3D): CirclePoint = apply(vector)
 	
 	
 	// OTHER    ----------------------------
@@ -27,14 +40,6 @@ object CirclePoint
 	 */
 	def apply(surfacePoint: CircleSurfacePoint, altitude: Distance): CirclePoint =
 		new _CirclePoint(surfacePoint, altitude)
-	
-	/**
-	 * @param latLong Latitude-longitude coordinate
-	 * @param altitude Altitude above the sea level
-	 * @return A 3D point matching those coordinates
-	 */
-	def apply(latLong: LatLong, altitude: Distance): CirclePoint =
-		new _CirclePoint(CircleSurfacePoint(latLong), altitude)
 	
 	/**
 	 * @param vector A 3D vector on the Circle of Earth -system
@@ -64,7 +69,7 @@ object CirclePoint
 		// ATTRIBUTES   ---------------------
 		
 		override lazy val toSurfacePoint: CircleSurfacePoint = CircleSurfacePoint(vector.in2D)
-		override lazy val altitude: Distance = CircleOfEarth.vectorLengthToDistance(vector.z)
+		override lazy val altitude: Distance = CircleOfEarth.distanceOf(vector.z)
 		
 		
 		// COMPUTED ------------------------
@@ -77,7 +82,7 @@ object CirclePoint
  * Represents a 3-dimensional point on the "Circle of Earth".
  *
  * The (0,0,0) vector lies at the "real" north pole, under the Polaris at sea level.
- * Unit vector length (1.0) represents the distance from the north pole to the equator.
+ * Length of 100 000 represents the distance from the north pole to the equator.
  * The z-axis runs parallel to the Earth's surface, towards the Heaven.
  * @author Mikko Hilpinen
  * @since 29.8.2023, v1.0